Testing a UI with no 'hard card' structure.
