--- src/components/TablatureGrill/TablatureGrill.test.js
@@ -1,9 +1,0 @@
-import TablatureGrill from "./TablatureGrill";
-import renderer from 'react-test-renderer';
-
-test('renders', ()=> {
-  const tree = renderer
-    .create(<TablatureGrill />)
-    .toJSON();
-  expect(tree).toMatchSnapshot();
-});
